Refactor web API client
